requirements.txt
asgiref==3.5.2
Django==4.1.3
django-betterforms==2.0.0
django-colorfield==0.8.0
django-extra-views==0.14.0
django-filter==22.1
django-search-views==0.3.7
Pillow==9.3.0
python-http-client==3.3.7
sendgrid==6.9.7
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
